Remove `AttributeLintKind::IllFormedAttributeInput` variant
