+++ compiler/rustc_lint_defs/src/lib.rs
@@ -654,11 +654,6 @@ pub enum DeprecatedSinceKind {
 
 #[derive(Debug)]
 pub enum AttributeLintKind {
-    IllFormedAttributeInput {
-        suggestions: Vec<String>,
-        docs: Option<&'static str>,
-        help: Option<String>,
-    },
     EmptyAttribute {
         first_span: Span,
         attr_path: String,
